==> SAGE/6_Physical_measures/2_Height.Rmd <==
---
title: "Height"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/heightDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/heightDS.R')
```

# Data process

## South Africa

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506`|
|-|-|
| **Label** | `true height (cm)`|
| **Table name** |`southAfricaINDdata`|
| **Units** |`cm`|
| **Missings** |`997 = refused`<br/>`998 = not able`|
| **Description** |  |



```{r assign1, echo=F}
opal.assign.table.tibble(o, 'SAGE_SA_w1','SAGE.southAfricaINDdata',
                         variables=list('q2506'), missings = TRUE)
SAGE_SA_w1 <- opal.execute(o,'SAGE_SA_w1')
```

```{r local1, echo=F}
vbl <- SAGE_SA_w1$q2506
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_SA_w1, aes(q2506)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep " refuse" into 997`
* `recode 998 "not able" into 993`

**R script:**

```{r harmo1}
SAGE_ds_SA_w1 <- tibble(id=SAGE_SA_w1$id)
SAGE_ds_SA_w1$height <- car::recode(SAGE_SA_w1$q2506, "NA=999; 998=993")
SAGE_ds_SA_w1$height <- labelled(SAGE_ds_SA_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vbl <- SAGE_ds_SA_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_SA_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")  
```

#### Validation

```{r crosstabulation1, echo=F}
kable(SAGE_SA_w1[which(SAGE_SA_w1$q2506<100),],caption = 'There are unusual values that are lower than 100cm:')
```







## Ghana

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506`|
|-|-|
| **Label** | `true height (cm)`|
| **Table name** |`ghanaINDData`|
| **Units** |`cm`|
| **Missings** |`997 = refused`<br/>`998 = not able`|
| **Description** |  |



```{r assign_ghana, echo=F}
opal.assign.table.tibble(o, 'SAGE_ghana_w1','SAGE.ghanaINDData',
                         variables=list('q2506'), missings = TRUE)
SAGE_ghana_w1 <- opal.execute(o,'SAGE_ghana_w1')
```

```{r local_ghana, echo=F}
vbl <- SAGE_ghana_w1$q2506
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_ghana_w1, aes(q2506)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep " refuse" into 997`
* `recode 998 "not able" into 993`


**R script:**

```{r harmo_ghana}
SAGE_ds_ghana_w1 <- tibble(id=SAGE_ghana_w1$id)
SAGE_ds_ghana_w1$height <- car::recode(SAGE_ghana_w1$q2506, "NA=999; 998=993")
SAGE_ds_ghana_w1$height <- labelled(SAGE_ds_ghana_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_ghana, echo=F}
vbl <- SAGE_ds_ghana_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_ghana_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")  
```

#### Validation

```{r crosstabulation_ghana, echo=F}
kable(SAGE_ghana_w1[which(SAGE_ghana_w1$q2506<100),], caption = 'There are unusual values that are lower than 100:')
```




## China

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506`|
|-|-|
| **Label** | `true height (cm)`|
| **Table name** |`chinaINDData`|
| **Units** |`cm`|
| **Missings** |`997 = refused`<br/>`998 = not able`|
| **Description** |  |



```{r assign_china, echo=F}
opal.assign.table.tibble(o, 'SAGE_china_w1','SAGE.chinaINDData',
                         variables=list('q2506'), missings = TRUE)
SAGE_china_w1 <- opal.execute(o,'SAGE_china_w1')
```

```{r local_china, echo=F}
vbl <- SAGE_china_w1$q2506
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_china_w1, aes(q2506)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```


#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep " refuse" into 997`
* `recode 998 "not able" into 993`


**R script:**

```{r harmo_china}
SAGE_ds_china_w1 <- tibble(id=SAGE_china_w1$id)
SAGE_ds_china_w1$height <- car::recode(SAGE_china_w1$q2506, "NA=999; 998=993")
SAGE_ds_china_w1$height <- labelled(SAGE_ds_china_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_china, echo=F}
vbl <- SAGE_ds_china_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_china_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")  
```

#### Validation

```{r crosstabulation_china, echo=F}
kable(SAGE_china_w1[which(SAGE_china_w1$q2506<100),], caption = 'There are unusual values that are lower than 100')
```









## India

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506`|
|-|-|
| **Label** | `true height (cm)`|
| **Table name** |`indiaINDData`|
| **Units** |`cm`|
| **Missings** |`997 = refused`<br/>`998 = not able`|
| **Description** |  |



```{r assign_india, echo=F}
opal.assign.table.tibble(o, 'SAGE_india_w1','SAGE.indiaINDData',
                         variables=list('q2506'), missings = TRUE)
SAGE_india_w1 <- opal.execute(o,'SAGE_india_w1')
```

```{r local_india, echo=F}
vbl <- SAGE_india_w1$q2506
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_india_w1, aes(q2506)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep " refuse" into 997`
* `recode 998 "not able" into 993`

**R script:**

```{r harmo_india}
SAGE_ds_india_w1 <- tibble(id=SAGE_india_w1$id)
SAGE_ds_india_w1$height <- car::recode(SAGE_india_w1$q2506, "NA=999; 998=993")
SAGE_ds_india_w1$height <- labelled(SAGE_ds_india_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_india, echo=F}
vbl <- SAGE_ds_india_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_india_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")  
```

#### Validation

```{r crosstabulation_india, echo=F}
kable(SAGE_india_w1[which(SAGE_india_w1$q2506<100),], caption = 'There are unusual values that are lower than 100. ')
```








## Russia

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506`|
|-|-|
| **Label** | `true height (cm)`|
| **Table name** |`russiaINDdata`|
| **Units** |`cm`|
| **Missings** |`997 = refused`<br/>`998 = not able`|
| **Description** |  |




```{r assign_russia, echo=F}
opal.assign.table.tibble(o, 'SAGE_russia_w1','SAGE.russiaINDdata',
                         variables=list('q2506'), missings = TRUE)
SAGE_russia_w1 <- opal.execute(o,'SAGE_russia_w1')
```

```{r local_russia, echo=F}
vbl <- SAGE_russia_w1$q2506
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_russia_w1, aes(q2506)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep " refuse" into 997`
* `recode 998 "not able" into 993`

**R script:**

```{r harmo_russia}
SAGE_ds_russia_w1 <- tibble(id=SAGE_russia_w1$id)
SAGE_ds_russia_w1$height <- car::recode(SAGE_russia_w1$q2506, "NA=999; 998=993")
SAGE_ds_russia_w1$height <- labelled(SAGE_ds_russia_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_russia, echo=F}
vbl <- SAGE_ds_russia_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_russia_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")  
```

#### Validation

```{r crosstabulation_russia, echo=F}
kable(SAGE_russia_w1[which(SAGE_russia_w1$q2506<100),], caption = 'There are unusual values that are lower than 100:')
```






## Mexico

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506`|
|-|-|
| **Label** | `true height (cm)`|
| **Table name** |`mexicoINDdata`|
| **Units** |`cm`|
| **Missings** |`997 = refused`<br/>`998 = not able`|
| **Description** |  |



```{r assign_Mexico, echo=F}
opal.assign.table.tibble(o, 'SAGE_mexico_w1','SAGE.mexicoINDdata',
                         variables=list('q2506'), missings = TRUE)
SAGE_mexico_w1 <- opal.execute(o,'SAGE_mexico_w1')
```

```{r local_Mexico, echo=F}
vbl <- SAGE_mexico_w1$q2506
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_mexico_w1, aes(q2506)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep " refuse" into 997`
* `recode 998 "not able" into 993`

**R script:**

```{r harmo_Mexico}
SAGE_ds_mexico_w1 <- tibble(id=SAGE_mexico_w1$id)
SAGE_ds_mexico_w1$height <- car::recode(SAGE_mexico_w1$q2506, "NA=999; 998=993")
SAGE_ds_mexico_w1$height <- labelled(SAGE_ds_mexico_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996,  "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_Mexico, echo=F}
vbl <- SAGE_ds_mexico_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_mexico_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")  
```

#### Validation
```{r crosstabulation_Mexico, echo=F}
```







## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with continuous data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for continuous data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of column with summary table categories
  t.summ <- summary(m.hds[[1]][2])[1:6]
  # Adding of missing/no-missing values categories
  t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    # First, summary values
    t.summ <- summary(m.hds[[i]][2])[1:6]
    # Next, missing/no-missing values
    t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
  }
  # Add sample size for each wave/population
  t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))
  # Add wave/population names
  dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


sage.cw <- list(SA_w1 = c("SA_w1","South Africa w1"), ghana_w1 = c("ghana_w1","Ghana w1"), china_w1 = c("china_w1","China w1"), india_w1 = c("india_w1","India w1"), russia_w1 = c("russia_w1","Russian Federation w1"), mexico_w1 = c("mexico_w1","Mexico w1"))

# Consider only harmonised waves
l.hds <- lapply(sage.cw, function(wname) if(exists(paste0("SAGE_ds_",wname[1]))){wname = list(get(paste0("SAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```






# Quality estimation
No comments.







```{r closeRsession, echo=FALSE}
opal.logout(o)
```
