==> SAGE/6_Physical_measures/3_BMI.Rmd <==
---
title: "BMI"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/bmiDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/bmiDS.R')
```


# Data process


**Functions:**

```{r function}


compute_bmi <- function(Heigth, Weigth, miss, heigth_m = TRUE){
  Heigth <- car::recode(Heigth, "NA=999")
  Weigth <- car::recode(Weigth, "NA=999")
  
  if(heigth_m){
    bmi <- Weigth/(Heigth^2)
  }else{
    bmi <- Weigth/((Heigth/100)^2)
  }
  for(msv in miss){
    bmi[Heigth==msv] <- msv
    bmi[Weigth==msv] <- msv
  }
  bmi
}

```



## South Africa

### Wave 1 

#### Study-specific variable description

| **Name** |`weight`|`height`|
|-|-|-|
| **Label** |`Weight of the participant in kg`|`Height of the participant`|
| **Table name** |`sage_SA_w1_hd`|`sage_SA_w1_hd`|
| **Units** |`kg`|`cm`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |  |

```{r assign SA_w1, echo=F}

load(paste0(datafolder,"SA_w1/weight",".RData"))
load(paste0(datafolder,"SA_w1/height",".RData"))

SAGE_SA_w1 <- left_join(weight,height,by="id")

rm(weight)
rm(height)

```

```{r local SA_w1, echo=F}

vbl <- SAGE_SA_w1$weight
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "weight")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_SA_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Weight") + ylab("Frequency") 

vbl <- SAGE_SA_w1$height
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "height")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_SA_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Height") + ylab("Frequency") 

```

#### Harmonisation algorithm
The harmonised variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `keep missing values into themselves`

**R script:**

```{r harmo SA_w1}


SAGE_SA_w1$bmi <- compute_bmi(Heigth = SAGE_SA_w1$height, 
                        Weigth = SAGE_SA_w1$weight, miss = miss_values_vector, 
                        heigth_m = FALSE)

SAGE_ds_SA_w1 <- tibble(id=SAGE_SA_w1$id)

SAGE_ds_SA_w1$bmi <- SAGE_SA_w1$bmi



```

#### Statistics of the new harmonised variable
```{r descript SA_w1, echo=F}

vbl <- SAGE_ds_SA_w1$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_SA_w1, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Validation

```{r crosstabulation SA_w1, echo=F}
kable(SAGE_SA_w1[which(SAGE_SA_w1$bmi>100 & SAGE_SA_w1$bmi<900),], caption = 'There are unusual values that are upper than 100. ')
```








## Ghana

### Wave 1 

#### Study-specific variable description

| **Name** |`weight`|`height`|
|-|-|-|
| **Label** |`Weight of the participant in kg`|`Height of the participant`|
| **Table name** |`sage_ghana_w1_hd`|`sage_ghana_w1_hd`|
| **Units** |`kg`|`cm`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |  |

```{r assign ghana_w1, echo=F}

load(paste0(datafolder,"ghana_w1/weight",".RData"))
load(paste0(datafolder,"ghana_w1/height",".RData"))

SAGE_ghana_w1 <- left_join(weight,height,by="id")

rm(weight)
rm(height)

```

```{r local ghana_w1, echo=F}

vbl <- SAGE_ghana_w1$weight
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "weight")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ghana_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Weight") + ylab("Frequency") 

vbl <- SAGE_ghana_w1$height
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "height")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ghana_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Height") + ylab("Frequency") 

```

#### Harmonisation algorithm
The harmonised variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `keep missing values into themselves`

**R script:**

```{r harmo ghana_w1}


SAGE_ghana_w1$bmi <- compute_bmi(Heigth = SAGE_ghana_w1$height, 
                        Weigth = SAGE_ghana_w1$weight, miss = miss_values_vector, 
                        heigth_m = FALSE)

SAGE_ds_ghana_w1 <- tibble(id=SAGE_ghana_w1$id)

SAGE_ds_ghana_w1$bmi <- SAGE_ghana_w1$bmi



```

#### Statistics of the new harmonised variable
```{r descript ghana_w1, echo=F}

vbl <- SAGE_ds_ghana_w1$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_ghana_w1, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Validation

```{r crosstabulation ghana_w1, echo=F}
kable(SAGE_ghana_w1[which(SAGE_ghana_w1$bmi>100 & SAGE_ghana_w1$bmi<900),], caption = 'There are unusual values that are higher than 100. ')
```




## China

### Wave 1 

#### Study-specific variable description

| **Name** |`weight`|`height`|
|-|-|-|
| **Label** |`Weight of the participant in kg`|`Height of the participant`|
| **Table name** |`sage_china_w1_hd`|`sage_china_w1_hd`|
| **Units** |`kg`|`cm`|
| **Missings** |`999 = Missing`|`999 = Missing`|
| **Description** |  |  |

```{r assign china_w1, echo=F}

load(paste0(datafolder,"china_w1/weight",".RData"))
load(paste0(datafolder,"china_w1/height",".RData"))

SAGE_china_w1 <- left_join(weight,height,by="id")

rm(weight)
rm(height)

```

```{r local china_w1, echo=F}

vbl <- SAGE_china_w1$weight
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "weight")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_china_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Weight") + ylab("Frequency") 

vbl <- SAGE_china_w1$height
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "height")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_china_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Height") + ylab("Frequency") 

```

#### Harmonisation algorithm
The harmonised variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `keep missing values into themselves`

**R script:**

```{r harmo china_w1}


SAGE_china_w1$bmi <- compute_bmi(Heigth = SAGE_china_w1$height, 
                        Weigth = SAGE_china_w1$weight, miss = miss_values_vector, 
                        heigth_m = FALSE)

SAGE_ds_china_w1 <- tibble(id=SAGE_china_w1$id)

SAGE_ds_china_w1$bmi <- SAGE_china_w1$bmi



```

#### Statistics of the new harmonised variable
```{r descript china_w1, echo=F}

vbl <- SAGE_ds_china_w1$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_china_w1, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Validation

```{r crosstabulation china_w1, echo=F}
kable(SAGE_china_w1[which((SAGE_china_w1$bmi>100 & SAGE_china_w1$bmi<900) | SAGE_china_w1$bmi<12),], caption = 'There are unusual values that are higher than 100 or lower than 12.')
```






## India

### Wave 1 

#### Study-specific variable description

| **Name** |`weight`|`height`|
|-|-|-|
| **Label** |`Weight of the participant in kg`|`Height of the participant`|
| **Table name** |`sage_india_w1_hd`|`sage_india_w1_hd`|
| **Units** |`kg`|`cm`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |  |

```{r assign india_w1, echo=F}

load(paste0(datafolder,"india_w1/weight",".RData"))
load(paste0(datafolder,"india_w1/height",".RData"))

SAGE_india_w1 <- left_join(weight,height,by="id")

rm(weight)
rm(height)

```

```{r local india_w1, echo=F}

vbl <- SAGE_india_w1$weight
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "weight")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_india_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Weight") + ylab("Frequency") 

vbl <- SAGE_india_w1$height
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "height")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_india_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Height") + ylab("Frequency") 

```

#### Harmonisation algorithm
The harmonised variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `keep missing values into themselves`

**R script:**

```{r harmo india_w1}


SAGE_india_w1$bmi <- compute_bmi(Heigth = SAGE_india_w1$height, 
                        Weigth = SAGE_india_w1$weight, miss = miss_values_vector, 
                        heigth_m = FALSE)

SAGE_ds_india_w1 <- tibble(id=SAGE_india_w1$id)

SAGE_ds_india_w1$bmi <- SAGE_india_w1$bmi



```

#### Statistics of the new harmonised variable
```{r descript india_w1, echo=F}

vbl <- SAGE_ds_india_w1$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_india_w1, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Validation

```{r crosstabulation india_w1, echo=F}
kable(SAGE_india_w1[which((SAGE_india_w1$bmi>100 & SAGE_india_w1$bmi<900) | SAGE_india_w1$bmi<12),], caption = 'There are unusual values that are higher than 100 or lower than 12.')
```









## Russia

### Wave 1 

#### Study-specific variable description

| **Name** |`weight`|`height`|
|-|-|-|
| **Label** |`Weight of the participant in kg`|`Height of the participant`|
| **Table name** |`sage_russia_w1_hd`|`sage_russia_w1_hd`|
| **Units** |`kg`|`cm`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |  |

```{r assign russia_w1, echo=F}

load(paste0(datafolder,"russia_w1/weight",".RData"))
load(paste0(datafolder,"russia_w1/height",".RData"))

SAGE_russia_w1 <- left_join(weight,height,by="id")

rm(weight)
rm(height)

```

```{r local russia_w1, echo=F}

vbl <- SAGE_russia_w1$weight
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "weight")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_russia_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Weight") + ylab("Frequency") 

vbl <- SAGE_russia_w1$height
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "height")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_russia_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Height") + ylab("Frequency") 

```

#### Harmonisation algorithm
The harmonised variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `keep missing values into themselves`

**R script:**

```{r harmo russia_w1}


SAGE_russia_w1$bmi <- compute_bmi(Heigth = SAGE_russia_w1$height, 
                        Weigth = SAGE_russia_w1$weight, miss = miss_values_vector, 
                        heigth_m = FALSE)

SAGE_ds_russia_w1 <- tibble(id=SAGE_russia_w1$id)

SAGE_ds_russia_w1$bmi <- SAGE_russia_w1$bmi



```

#### Statistics of the new harmonised variable
```{r descript russia_w1, echo=F}

vbl <- SAGE_ds_russia_w1$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_russia_w1, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Validation

```{r crosstabulation russia_w1, echo=F}
kable(SAGE_russia_w1[which((SAGE_russia_w1$bmi>100 & SAGE_russia_w1$bmi<900) | SAGE_russia_w1$bmi<12),], caption = 'There are unusual values that are higher than 100 or lower than 12.')
```





## Mexico

### Wave 1 

#### Study-specific variable description

| **Name** |`weight`|`height`|
|-|-|-|
| **Label** |`Weight of the participant in kg`|`Height of the participant`|
| **Table name** |`sage_mexico_w1_hd`|`sage_mexico_w1_hd`|
| **Units** |`kg`|`cm`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |  |

```{r assign mexico_w1, echo=F}

load(paste0(datafolder,"mexico_w1/weight",".RData"))
load(paste0(datafolder,"mexico_w1/height",".RData"))

SAGE_mexico_w1 <- left_join(weight,height,by="id")

rm(weight)
rm(height)

```

```{r local mexico_w1, echo=F}

vbl <- SAGE_mexico_w1$weight
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "weight")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_mexico_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Weight") + ylab("Frequency") 

vbl <- SAGE_mexico_w1$height
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[3], caption = "height")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_mexico_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("Height") + ylab("Frequency") 

```

#### Harmonisation algorithm
The harmonised variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `keep missing values into themselves`

**R script:**

```{r harmo mexico_w1}


SAGE_mexico_w1$bmi <- compute_bmi(Heigth = SAGE_mexico_w1$height, 
                        Weigth = SAGE_mexico_w1$weight, miss = miss_values_vector, 
                        heigth_m = FALSE)

SAGE_ds_mexico_w1 <- tibble(id=SAGE_mexico_w1$id)

SAGE_ds_mexico_w1$bmi <- SAGE_mexico_w1$bmi



```

#### Statistics of the new harmonised variable
```{r descript mexico_w1, echo=F}

vbl <- SAGE_ds_mexico_w1$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_mexico_w1, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency") 

```

#### Validation

```{r crosstabulation mexico_w1, echo=F}

```












## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with continuous data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for continuous data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of column with summary table categories
  t.summ <- summary(m.hds[[1]][2])[1:6]
  # Adding of missing/no-missing values categories
  t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    # First, summary values
    t.summ <- summary(m.hds[[i]][2])[1:6]
    # Next, missing/no-missing values
    t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
  }
  # Add sample size for each wave/population
  t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))
  # Add wave/population names
  dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


sage.cw <- list(SA_w1 = c("SA_w1","South Africa w1"), ghana_w1 = c("ghana_w1","Ghana w1"), china_w1 = c("china_w1","China w1"), india_w1 = c("india_w1","India w1"), russia_w1 = c("russia_w1","Russian Federation w1"), mexico_w1 = c("mexico_w1","Mexico w1"))

# Consider only harmonised waves
l.hds <- lapply(sage.cw, function(wname) if(exists(paste0("SAGE_ds_",wname[1]))){wname = list(get(paste0("SAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```




# Quality estimation

Individuals with implaussible data (`bmi>100` or `bmi<10`) have been specified at each validation with their harmonised weights and heights (which coincide with the study-specific variables `q2507` and `q2506` up to some missing values).

```{r closeRsession, echo=FALSE}
opal.logout(o)
```
