==> SAGE/6_Physical_measures/4_Obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/obesityDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/obesityDS.R')
```

# Data process


## South Africa

### Wave 1

#### Study-specific variable description



| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`sage_SA_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |





```{r assign SA_w1, include=FALSE}

load(paste0(datafolder,"SA_w1/bmi",".RData"))
SAGE_bmi <- bmi
SAGE_bmi$bmi <- as.numeric(SAGE_bmi$bmi)
rm(bmi)

```

```{r local SA_w1, echo=F}

vbl <- SAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo SA_w1}

SAGE_ds_SA_w1 <- tibble(id=SAGE_bmi$id)
SAGE_ds_SA_w1$obesity <- car::recode(SAGE_bmi[[2]], "30:990=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript SA_w1, echo=F}

vari <- SAGE_ds_SA_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SAGE_ds_SA_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")

```

#### Validation
```{r crosstab SA_w1, echo=F}
BeforeH <- c(sum(na.omit(SAGE_bmi[[2]])<30),sum(na.omit(SAGE_bmi[[2]])>=30 & na.omit(SAGE_bmi[[2]])<=900), sum(na.omit(SAGE_bmi[[2]])==993), sum(na.omit(SAGE_bmi[[2]])==997), sum(na.omit(SAGE_bmi[[2]])==999))
AfterH <- table(SAGE_ds_SA_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","993", "997", "999") 
kable(C)
```


```{r read SA_w1, include=FALSE}
rm(SAGE_bmi)
```


## Ghana

### Wave 1

#### Study-specific variable description



| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`sage_ghana_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |

    
```{r assign2, include=FALSE}

load(paste0(datafolder,"ghana_w1/bmi",".RData"))
SAGE_bmi <- bmi
SAGE_bmi$bmi <- as.numeric(SAGE_bmi$bmi)
rm(bmi)

```

```{r local2, echo=F}

vbl <- SAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo2}

SAGE_ds_ghana_w1 <- tibble(id=SAGE_bmi$id)
SAGE_ds_ghana_w1$obesity <- car::recode(SAGE_bmi[[2]], "30:990=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript2, echo=F}

vari <- SAGE_ds_ghana_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SAGE_ds_ghana_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")

```

#### Validation
```{r crosstab2, echo=F}
BeforeH <- c(sum(na.omit(SAGE_bmi[[2]])<30),sum(na.omit(SAGE_bmi[[2]])>=30 & na.omit(SAGE_bmi[[2]])<=900), sum(na.omit(SAGE_bmi[[2]])==993), sum(na.omit(SAGE_bmi[[2]])==997), sum(na.omit(SAGE_bmi[[2]])==999))
AfterH <- table(SAGE_ds_ghana_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","993", "997", "999") 
kable(C)
```



```{r read2, include=FALSE}
rm(SAGE_bmi)
```




## China

### Wave 1

#### Study-specific variable description




| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`sage_china_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`999 = Missing`|
| **Description** |  |

    
```{r assign4, include=FALSE}

load(paste0(datafolder,"china_w1/bmi",".RData"))
SAGE_bmi <- bmi
SAGE_bmi$bmi <- as.numeric(SAGE_bmi$bmi)
rm(bmi)

```

```{r local4, echo=F}

vbl <- SAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo4}

SAGE_ds_china_w1 <- tibble(id=SAGE_bmi$id)
SAGE_ds_china_w1$obesity <- car::recode(SAGE_bmi[[2]], "30:990=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript4, echo=F}

vari <- SAGE_ds_china_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SAGE_ds_china_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")

```

#### Validation
```{r crosstab4, echo=F}
BeforeH <- c(sum(na.omit(SAGE_bmi[[2]])<30),sum(na.omit(SAGE_bmi[[2]])>=30 & na.omit(SAGE_bmi[[2]])<=900), sum(na.omit(SAGE_bmi[[2]])==999))
AfterH <- table(SAGE_ds_china_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1", "999") 
kable(C)
```



```{r read4, include=FALSE}
rm(SAGE_bmi)
```



## India

### Wave 1

#### Study-specific variable description




| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`sage_india_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |

    
```{r assign5, include=FALSE}

load(paste0(datafolder,"india_w1/bmi",".RData"))
SAGE_bmi <- bmi
SAGE_bmi$bmi <- as.numeric(SAGE_bmi$bmi)
rm(bmi)

```

```{r local5, echo=F}

vbl <- SAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo5}

SAGE_ds_india_w1 <- tibble(id=SAGE_bmi$id)
SAGE_ds_india_w1$obesity <- car::recode(SAGE_bmi[[2]], "30:990=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript5, echo=F}

vari <- SAGE_ds_india_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SAGE_ds_india_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")

```

#### Validation
```{r crosstab5, echo=F}
BeforeH <- c(sum(na.omit(SAGE_bmi[[2]])<30),sum(na.omit(SAGE_bmi[[2]])>=30 & na.omit(SAGE_bmi[[2]])<=900), sum(na.omit(SAGE_bmi[[2]])==993), sum(na.omit(SAGE_bmi[[2]])==997), sum(na.omit(SAGE_bmi[[2]])==999))
AfterH <- table(SAGE_ds_india_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","993", "997", "999") 
kable(C)
```



```{r read5, include=FALSE}
rm(SAGE_bmi)
```


## Russia

### Wave 1

#### Study-specific variable description


| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`sage_russia_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |

    
```{r assign6, include=FALSE}

load(paste0(datafolder,"russia_w1/bmi",".RData"))
SAGE_bmi <- bmi
SAGE_bmi$bmi <- as.numeric(SAGE_bmi$bmi)
rm(bmi)

```

```{r local6, echo=F}

vbl <- SAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo6}

SAGE_ds_russia_w1 <- tibble(id=SAGE_bmi$id)
SAGE_ds_russia_w1$obesity <- car::recode(SAGE_bmi[[2]], "30:990=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript6, echo=F}

vari <- SAGE_ds_russia_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SAGE_ds_russia_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")

```

#### Validation
```{r crosstab6, echo=F}
BeforeH <- c(sum(na.omit(SAGE_bmi[[2]])<30),sum(na.omit(SAGE_bmi[[2]])>=30 & na.omit(SAGE_bmi[[2]])<=900), sum(na.omit(SAGE_bmi[[2]])==993), sum(na.omit(SAGE_bmi[[2]])==997), sum(na.omit(SAGE_bmi[[2]])==999))
AfterH <- table(SAGE_ds_russia_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","993", "997", "999") 
kable(C)
```



```{r read6, include=FALSE}
rm(SAGE_bmi)
```


## Mexico

### Wave 1

#### Study-specific variable description




| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`sage_mexico_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |

    
```{r assign7, include=FALSE}

load(paste0(datafolder,"mexico_w1/bmi",".RData"))
SAGE_bmi <- bmi
SAGE_bmi$bmi <- as.numeric(SAGE_bmi$bmi)
rm(bmi)

```

```{r local7, echo=F}

vbl <- SAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo7}

SAGE_ds_mexico_w1 <- tibble(id=SAGE_bmi$id)
SAGE_ds_mexico_w1$obesity <- car::recode(SAGE_bmi[[2]], "30:990=1; lo:30=0; NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript7, echo=F}

vari <- SAGE_ds_mexico_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SAGE_ds_mexico_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")

```

#### Validation
```{r crosstab7, echo=F}
BeforeH <- c(sum(na.omit(SAGE_bmi[[2]])<30),sum(na.omit(SAGE_bmi[[2]])>=30 & na.omit(SAGE_bmi[[2]])<=900), sum(na.omit(SAGE_bmi[[2]])==993), sum(na.omit(SAGE_bmi[[2]])==997), sum(na.omit(SAGE_bmi[[2]])==999))
AfterH <- table(SAGE_ds_mexico_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","993", "997", "999") 
kable(C)
```




## Summary of descriptive statistics of the harmonised variable accross populations and waves




```{r helpfunctions, echo=F}


labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with categorical data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cat_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Labelling of categories (for categorical variables, only non-missing values)
    aux_object[[2]] <- labelled(aux_object[[2]], labels = cat_label[1:2])
  # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for categorical data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of columns with categories and labels
  t.hds <- frq(l.hds[[1]][2])[[1]][,c(1,2)] 
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
  }
  # Add sample size for each wave/population
  t.hds[13,] <- c("n", "sample size", sapply(l.hds,function(wave) length(wave[[1]]))
  )
  # Add wave/population names
  names(t.hds)<-c("val", "label",lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


sage.cw <- list(SA_w1 = c("SA_w1","South Africa w1"), ghana_w1 = c("ghana_w1","Ghana w1"), china_w1 = c("china_w1","China w1"), india_w1 = c("india_w1","India w1"), russia_w1 = c("russia_w1","Russian Federation w1"), mexico_w1 = c("mexico_w1","Mexico w1"))

# Consider only harmonised waves
l.hds <- lapply(sage.cw, function(wname) if(exists(paste0("SAGE_ds_",wname[1]))){wname = list(get(paste0("SAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```



# Quality estimation

Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.


```{r read7, include=FALSE}
rm(SAGE_bmi)
```




<!--- ## Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
