==> SAGE/6_Physical_measures/7_Grip_Strength.Rmd <==
---
title: "Grip strength"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `grip`
* Variable label: `Grip strength in kg`
* Variable description: `Grip strength`
* Domain: `Physical measures`
* Value type: `continuous`

* Comments: `According to a systematic review (Roberts et al, 2011), there are some inconsistencies in the use of maximum or mean grip strength as a summary measure which limit comparison of results between epidemiological studies. With multiple attempts, the maximum grip strength will be greater than the mean value. 
Even though for right handed people, there is a 10% stronger grip in the dominant hand, this is not true for left handed or ambidextrous people, whose grip strength are more or less equal in both hands (Roberts et al, 2011). 
In light of the above, the harmonised variable "grip" is proposed to be the highest score (to the nearest kg) of all hand grip measurements in each study, regardless of hand dominance. This is following a protocol suggested by Roberts et al (2011), based on the American Society of Hand Therapists (ASHT) protocol.`

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Grip strength"
```

# Data process

## South Africa

### Wave 1 

#### Study-specific variable description

| **Name** | `q2521`| `q2522`| `q2523`| `q2524`|
|-|-|-|-|-|
| **Label** | `q2521: left 1` | `q2522: left 2` | `q2523: right 1` | `q2524: right 2` |
| **Table name** | `southAfricaINDdata` |`southAfricaINDdata` |`southAfricaINDdata` |`southAfricaINDdata` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing categories** | `-9 = refused` |  `-9 = refused` | `-9 = refused` | `-9 = refused` |
| **Description** |  | | | |

```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'SAGE_gs','SAGE.southAfricaINDdata',
                         variables=list('q2521', 'q2522', 'q2523', 'q2524'), missings = TRUE)
SAGE_gs <- opal.execute(o,'SAGE_gs')
```

```{r local1, echo=FALSE}
vari <- SAGE_gs$q2521
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2521)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2522
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2522)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2523
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2523)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2524
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2524)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `NA into 999`

**R script:**

```{r harmo1}
SAGE_SA_ds_w1 <- tibble(id=SAGE_gs$id)
SAGE_SA_ds_w1$grip <- apply(SAGE_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
SAGE_SA_ds_w1$grip <- car::recode(SAGE_SA_ds_w1$grip, "-Inf=999; -9=997")
SAGE_SA_ds_w1$grip <- as.integer(SAGE_SA_ds_w1$grip)
SAGE_SA_ds_w1$grip <- labelled(SAGE_SA_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- SAGE_SA_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_SA_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```

```{r importOPAL1, echo=FALSE}
#val_labels(SAGE_SA_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_SA_ds_w1', SAGE_SA_ds_w1)
#opal.symbol_import(o,'SAGE_SA_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_SA_ds_w1$grip) <- ds_label
grip <- SAGE_SA_ds_w1
save(grip, file = "../RData/SA_w1/grip.RData")
rm(grip, SAGE_grip)
```




## Ghana

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `q2521`| `q2522`| `q2523`| `q2524`|
|-|-|-|-|-|
| **Label** | `q2521: left 1` | `q2522: left 2` | `q2523: right 1` | `q2524: right 2` |
| **Table name** | `ghanaINDData` |`ghanaINDData` |`ghanaINDData` |`ghanaINDData` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing categories** | `-9 = refused` |  `-9 = refused` | `-9 = refused` | `-9 = refused` |
| **Description** |  | | | |


```{r assign_ghana, echo=FALSE}
opal.assign.table.tibble(o, 'SAGE_gs','SAGE.ghanaINDData',
                         variables=list('q2521', 'q2522', 'q2523', 'q2524'), missings = TRUE)
SAGE_gs <- opal.execute(o,'SAGE_gs')
```

```{r local_ghana, echo=FALSE}
vari <- SAGE_gs$q2521
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2521)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2522
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2522)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2523
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2523)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2524
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2524)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `NA into 999`

**R script:**

```{r harmo_ghana}
SAGE_ghana_ds_w1 <- tibble(id=SAGE_gs$id)
SAGE_ghana_ds_w1$grip <- apply(SAGE_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
SAGE_ghana_ds_w1$grip <- car::recode(SAGE_ghana_ds_w1$grip, "-Inf=999; -9=997")
SAGE_ghana_ds_w1$grip <- as.integer(SAGE_ghana_ds_w1$grip)
SAGE_ghana_ds_w1$grip <- labelled(SAGE_ghana_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_ghana, echo=FALSE}
vari <- SAGE_ghana_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ghana_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_ghana, echo=FALSE}
```

```{r importOPAL_ghana, echo=FALSE}
#val_labels(SAGE_ghana_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_ghana_ds_w1', SAGE_ghana_ds_w1)
#opal.symbol_import(o,'SAGE_ghana_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_ghana_ds_w1$grip) <- ds_label
grip <- SAGE_ghana_ds_w1
save(grip, file = "../RData/ghana_w1/grip.RData")
rm(grip)
```





## China

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `q2521`| `q2522`| `q2523`| `q2524`|
|-|-|-|-|-|
| **Label** | `q2521: left 1` | `q2522: left 2` | `q2523: right 1` | `q2524: right 2` |
| **Table name** | `chinaINDData` |`chinaINDData` |`chinaINDData` |`chinaINDData` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing categories** | `-9 = refused` |  `-9 = refused` | `-9 = refused` | `-9 = refused` |
| **Description** |  | | | |

```{r assign_china, echo=FALSE}
opal.assign.table.tibble(o, 'SAGE_gs','SAGE.chinaINDData',
                         variables=list('q2521', 'q2522', 'q2523', 'q2524'), missings = TRUE)
SAGE_gs <- opal.execute(o,'SAGE_gs')
```

```{r local_china, echo=FALSE}
vari <- SAGE_gs$q2521
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2521)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2522
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2522)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2523
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2523)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2524
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2524)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```


#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `NA into 999`

**R script:**

```{r harmo_china}
SAGE_china_ds_w1 <- tibble(id=SAGE_gs$id)
SAGE_china_ds_w1$grip <- apply(SAGE_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
SAGE_china_ds_w1$grip <- car::recode(SAGE_china_ds_w1$grip, "-Inf=999; -9=997")
SAGE_china_ds_w1$grip <- as.integer(SAGE_china_ds_w1$grip)
SAGE_china_ds_w1$grip <- labelled(SAGE_china_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_china, echo=FALSE}
vari <- SAGE_china_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_china_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_china, echo=FALSE}
```

```{r importOPAL_china, echo=FALSE}
#val_labels(SAGE_china_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_china_ds_w1', SAGE_china_ds_w1)
#opal.symbol_import(o,'SAGE_china_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_china_ds_w1$grip) <- ds_label
grip <- SAGE_china_ds_w1
save(grip, file = "../RData/china_w1/grip.RData")
rm(grip)
```





## India

### Wave 1 

| **Name** | `q2521`| `q2522`| `q2523`| `q2524`|
|-|-|-|-|-|
| **Label** | `q2521: left 1` | `q2522: left 2` | `q2523: right 1` | `q2524: right 2` |
| **Table name** | `indiaINDData` |`indiaINDData` |`indiaINDData` |`indiaINDData` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing categories** | `-9 = refused` |  `-9 = refused` | `-9 = refused` | `-9 = refused` |
| **Description** |  | | | |


```{r assign_india, echo=FALSE}
opal.assign.table.tibble(o, 'SAGE_gs','SAGE.indiaINDData',
                         variables=list('q2521', 'q2522', 'q2523', 'q2524'), missings = TRUE)
SAGE_gs <- opal.execute(o,'SAGE_gs')
```

```{r local_india, echo=FALSE}
vari <- SAGE_gs$q2521
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2521)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2522
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2522)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2523
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2523)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2524
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2524)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `NA into 999`

**R script:**

```{r harmo_india}
SAGE_india_ds_w1 <- tibble(id=SAGE_gs$id)
SAGE_india_ds_w1$grip <- apply(SAGE_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
SAGE_india_ds_w1$grip <- car::recode(SAGE_india_ds_w1$grip, "-Inf=999; -9=997")
SAGE_india_ds_w1$grip <- as.integer(SAGE_india_ds_w1$grip)
SAGE_india_ds_w1$grip <- labelled(SAGE_india_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_india, echo=FALSE}
vari <- SAGE_india_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_india_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_india, echo=FALSE}
```

```{r importOPAL_india, echo=FALSE}
#val_labels(SAGE_india_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_india_ds_w1', SAGE_india_ds_w1)
#opal.symbol_import(o,'SAGE_india_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_india_ds_w1$grip) <- ds_label
grip <- SAGE_india_ds_w1
save(grip, file = "../RData/india_w1/grip.RData")
rm(grip)
```






## Russia

### Wave 1 

#### Study-specific variable description

| **Name** | `q2521`| `q2522`| `q2523`| `q2524`|
|-|-|-|-|-|
| **Label** | `q2521: left 1` | `q2522: left 2` | `q2523: right 1` | `q2524: right 2` |
| **Table name** | `russiaINDdata` |`russiaINDdata` |`russiaINDdata` |`russiaINDdata` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing categories** | `-9 = refused` |  `-9 = refused` | `-9 = refused` | `-9 = refused` |
| **Description** |  | | | |


```{r assign_russia, echo=FALSE}
opal.assign.table.tibble(o, 'SAGE_gs','SAGE.russiaINDdata',
                         variables=list('q2521', 'q2522', 'q2523', 'q2524'), missings = TRUE)
SAGE_gs <- opal.execute(o,'SAGE_gs')
```

```{r local_russia, echo=FALSE}
vari <- SAGE_gs$q2521
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2521)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2522
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2522)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2523
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2523)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2524
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2524)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `NA into 999`

**R script:**

```{r harmo_russia}
SAGE_russia_ds_w1 <- tibble(id=SAGE_gs$id)
SAGE_russia_ds_w1$grip <- apply(SAGE_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
SAGE_russia_ds_w1$grip <- car::recode(SAGE_russia_ds_w1$grip, "-Inf=999; -9=997")
SAGE_russia_ds_w1$grip <- as.integer(SAGE_russia_ds_w1$grip)
SAGE_russia_ds_w1$grip <- labelled(SAGE_russia_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_russia, echo=FALSE}
vari <- SAGE_russia_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_russia_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_russia, echo=FALSE}
```

```{r importOPAL_russia, echo=FALSE}
#val_labels(SAGE_russia_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_russia_ds_w1', SAGE_russia_ds_w1)
#opal.symbol_import(o,'SAGE_russia_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_russia_ds_w1$grip) <- ds_label
grip <- SAGE_russia_ds_w1
save(grip, file = "../RData/russia_w1/grip.RData")
rm(grip)
```






## Mexico

### Wave 1 

#### Study-specific variable description

| **Name** | `q2521`| `q2522`| `q2523`| `q2524`|
|-|-|-|-|-|
| **Label** | `q2521: left 1` | `q2522: left 2` | `q2523: right 1` | `q2524: right 2` |
| **Table name** | `mexicoINDdata` |`mexicoINDdata` |`mexicoINDdata` |`mexicoINDdata` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing categories** | `-9 = refused` |  `-9 = refused` | `-9 = refused` | `-9 = refused` |
| **Description** |  | | | |


```{r assign_Mexico, echo=FALSE}
opal.assign.table.tibble(o, 'SAGE_gs','SAGE.mexicoINDdata',
                         variables=list('q2521', 'q2522', 'q2523', 'q2524'), missings = TRUE)
SAGE_gs <- opal.execute(o,'SAGE_gs')
```

```{r local_Mexico, echo=FALSE}
vari <- SAGE_gs$q2521
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2521)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2522
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2522)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2523
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2523)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 

vari <- SAGE_gs$q2524
kable(Continuous_summary(var = vari, missing_values = c(-9))[3])
kable(Continuous_summary(vari, missing_values = c(-9))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-9))$summary, caption = "Summary")
ggplot(SAGE_gs, aes(q2524)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `-9 into 997`
* `NA into 999`

**R script:**

```{r harmo_Mexico}
SAGE_mexico_ds_w1 <- tibble(id=SAGE_gs$id)
SAGE_mexico_ds_w1$grip <- apply(SAGE_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
SAGE_mexico_ds_w1$grip <- car::recode(SAGE_mexico_ds_w1$grip, "-Inf=999; -9=997")
SAGE_mexico_ds_w1$grip <- as.integer(SAGE_mexico_ds_w1$grip)
SAGE_mexico_ds_w1$grip <- labelled(SAGE_mexico_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_Mexico, echo=FALSE}
vari <- SAGE_mexico_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_mexico_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_Mexico, echo=FALSE}
```


```{r importOPAL_Mexico, echo=FALSE}
#val_labels(SAGE_mexico_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_mexico_ds_w1', SAGE_mexico_ds_w1)
#opal.symbol_import(o,'SAGE_mexico_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_mexico_ds_w1$grip) <- ds_label
grip <- SAGE_mexico_ds_w1
save(grip, file = "../RData/mexico_w1/grip.RData")
rm(grip)
```



# Quality control
No comments.





```{r closeRsession, echo=FALSE}
opal.logout(o)
```
