==> SAGE/6_Physical_measures/12_ah.Rmd <==
---
title: "Arterial hypertension"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `ah`
* Variable label: `Arterial hypertension`
* Variable description: `Arterial hypertension. Means of all diastolic blood pressure measures.`
* Domain: `Physical Measures`
* Value type: `Categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | no hypertension
1        | hypertension

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Arterial hypertension"
ds_label_all <- "Arterial hypertension. Means of all diastolic blood pressure measures."
```




# Data process

## South Africa

### Wave 1 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `ah` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign1, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/SA_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/SA_w1/dbp.RData")
SAGE_ah <- merge(sbp, dbp, by = "id")
```

```{r local1, echo=FALSE}
vari <- SAGE_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- SAGE_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo1}
SAGE_ah$dh <- car::recode(SAGE_ah$dbp, "0:89.99999999=0; 90:900=1")
SAGE_ah$sh <- car::recode(SAGE_ah$sbp, "0:139.9999999=0; 140:900=1")

SAGE_SA_ds_w1 <- tibble(id=SAGE_ah$id)
SAGE_SA_ds_w1$ah <- NA
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==0 & SAGE_ah$sh==0)] <- 0
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==1 & SAGE_ah$sh==0)] <- 1
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==0 & SAGE_ah$sh==1)] <- 1
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==1 & SAGE_ah$sh==1)] <- 1
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==1 & SAGE_ah$sh==998)] <- 1
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==998 & SAGE_ah$sh==1)] <- 1
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==998 & SAGE_ah$sh==0)] <- 998
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==999 & SAGE_ah$sh==999)] <- 999
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==998 & SAGE_ah$sh==998)] <- 998
SAGE_SA_ds_w1$ah[which(SAGE_ah$dh==997 & SAGE_ah$sh==997)] <- 997
SAGE_SA_ds_w1$ah <- labelled(SAGE_SA_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- SAGE_SA_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SAGE_SA_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```

```{r importOPAL1, echo=FALSE}
#val_labels(SAGE_SA_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_SA_ds_w1', SAGE_SA_ds_w1)
#opal.symbol_import(o,'SAGE_SA_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_SA_ds_w1$ah) <- ds_label_all
ah <- SAGE_SA_ds_w1
save(ah, file = "../RData/SA_w1/ah.RData")
rm(ah, SAGE_ah)
```




## Ghana

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `ah`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `ah` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign_ghana, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/ghana_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/ghana_w1/dbp.RData")
SAGE_g_ah <- merge(sbp, dbp, by = "id")
```

```{r local_ghana, echo=FALSE}
vari <- SAGE_g_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_g_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- SAGE_g_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_g_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_ghana}
SAGE_g_ah$dh <- car::recode(SAGE_g_ah$dbp, "0:89.99999999=0; 90:900=1")
SAGE_g_ah$sh <- car::recode(SAGE_g_ah$sbp, "0:139.9999999=0; 140:900=1")


SAGE_ghana_ds_w1 <- tibble(id=SAGE_g_ah$id)
SAGE_ghana_ds_w1$ah <- NA
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==0 & SAGE_g_ah$sh==0)] <- 0
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==1 & SAGE_g_ah$sh==0)] <- 1
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==0 & SAGE_g_ah$sh==1)] <- 1
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==1 & SAGE_g_ah$sh==1)] <- 1
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==999 & SAGE_g_ah$sh==999)] <- 999
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==998 & SAGE_g_ah$sh==998)] <- 998
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==997 & SAGE_g_ah$sh==997)] <- 997
SAGE_ghana_ds_w1$ah[which(SAGE_g_ah$dh==998 & SAGE_g_ah$sh==1)] <- 1
SAGE_ghana_ds_w1$ah <- labelled(SAGE_ghana_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_ghana, echo=FALSE}
vari <- SAGE_ghana_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SAGE_ghana_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_ghana, echo=FALSE}
```

```{r importOPAL_ghana, echo=FALSE}
#val_labels(SAGE_ghana_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_ghana_ds_w1', SAGE_ghana_ds_w1)
#opal.symbol_import(o,'SAGE_ghana_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_ghana_ds_w1$ah) <- ds_label_all
ah <- SAGE_ghana_ds_w1
save(ah, file = "../RData/ghana_w1/ah.RData")
rm(ah)
```





## China

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign_china, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/china_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/china_w1/dbp.RData")
SAGE_ch_ah <- merge(sbp, dbp, by = "id")
```

```{r local_china, echo=FALSE}
vari <- SAGE_ch_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ch_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- SAGE_ch_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ch_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```


#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_china}
SAGE_ch_ah$dh <- car::recode(SAGE_ch_ah$dbp, "0:89.99999999=0; 90:900=1")
SAGE_ch_ah$sh <- car::recode(SAGE_ch_ah$sbp, "0:139.9999999=0; 140:900=1")


SAGE_china_ds_w1 <- tibble(id=SAGE_ch_ah$id)
SAGE_china_ds_w1$ah <- NA
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==0 & SAGE_ch_ah$sh==0)] <- 0
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==1 & SAGE_ch_ah$sh==0)] <- 1
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==0 & SAGE_ch_ah$sh==1)] <- 1
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==1 & SAGE_ch_ah$sh==1)] <- 1
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==999 & SAGE_ch_ah$sh==999)] <- 999
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==998 & SAGE_ch_ah$sh==998)] <- 998
SAGE_china_ds_w1$ah[which(SAGE_ch_ah$dh==997 & SAGE_ch_ah$sh==997)] <- 997
SAGE_china_ds_w1$ah <- labelled(SAGE_china_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_china, echo=FALSE}
vari <- SAGE_china_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SAGE_china_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_china, echo=FALSE}
```

```{r importOPAL_china, echo=FALSE}
#val_labels(SAGE_china_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_china_ds_w1', SAGE_china_ds_w1)
#opal.symbol_import(o,'SAGE_china_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_china_ds_w1$ah) <- ds_label_all
ah <- SAGE_china_ds_w1
save(ah, file = "../RData/china_w1/ah.RData")
rm(ah)
```





## India

### Wave 1 

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign_india, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/india_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/india_w1/dbp.RData")
SAGE_ind_ah <- merge(sbp, dbp, by = "id")
```

```{r local_india, echo=FALSE}
vari <- SAGE_ind_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ind_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- SAGE_ind_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_ind_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_india}
SAGE_ind_ah$dh <- car::recode(SAGE_ind_ah$dbp, "0:89.99999999=0; 90:900=1")
SAGE_ind_ah$sh <- car::recode(SAGE_ind_ah$sbp, "0:139.9999999=0; 140:900=1")

SAGE_india_ds_w1 <- tibble(id=SAGE_ind_ah$id)
SAGE_india_ds_w1$ah <- NA
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==0 & SAGE_ind_ah$sh==0)] <- 0
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==1 & SAGE_ind_ah$sh==0)] <- 1
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==0 & SAGE_ind_ah$sh==1)] <- 1
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==1 & SAGE_ind_ah$sh==1)] <- 1
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==999 & SAGE_ind_ah$sh==999)] <- 999
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==998 & SAGE_ind_ah$sh==998)] <- 998
SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==997 & SAGE_ind_ah$sh==997)] <- 997

SAGE_india_ds_w1$ah[which(SAGE_ind_ah$dh==998 & SAGE_ind_ah$sh==0)] <- 998

SAGE_india_ds_w1$ah <- labelled(SAGE_india_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_india, echo=FALSE}
vari <- SAGE_india_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SAGE_india_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_india, echo=FALSE}
```

```{r importOPAL_india, echo=FALSE}
#val_labels(SAGE_india_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_india_ds_w1', SAGE_india_ds_w1)
#opal.symbol_import(o,'SAGE_india_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_india_ds_w1$ah) <- ds_label_all
ah <- SAGE_india_ds_w1
save(ah, file = "../RData/india_w1/ah.RData")
rm(ah)
```






## Russia

### Wave 1 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign_russia, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/russia_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/russia_w1/dbp.RData")
SAGE_rus_ah <- merge(sbp, dbp, by = "id")
```

```{r local_russia, echo=FALSE}
vari <- SAGE_rus_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_rus_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- SAGE_rus_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_rus_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_russia}
SAGE_rus_ah$dh <- car::recode(SAGE_rus_ah$dbp, "0:89.99999999=0; 90:900=1")
SAGE_rus_ah$sh <- car::recode(SAGE_rus_ah$sbp, "0:139.9999999=0; 140:900=1")

SAGE_russia_ds_w1 <- tibble(id=SAGE_rus_ah$id)
SAGE_russia_ds_w1$ah <- NA
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==0 & SAGE_rus_ah$sh==0)] <- 0
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==1 & SAGE_rus_ah$sh==0)] <- 1
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==0 & SAGE_rus_ah$sh==1)] <- 1
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==1 & SAGE_rus_ah$sh==1)] <- 1
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==999 & SAGE_rus_ah$sh==999)] <- 999
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==998 & SAGE_rus_ah$sh==998)] <- 998
SAGE_russia_ds_w1$ah[which(SAGE_rus_ah$dh==997 & SAGE_rus_ah$sh==997)] <- 997

SAGE_russia_ds_w1$ah <- labelled(SAGE_russia_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_russia, echo=FALSE}
vari <- SAGE_russia_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SAGE_russia_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_russia, echo=FALSE}
```


```{r importOPAL_russia, echo=FALSE}
#val_labels(SAGE_russia_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_russia_ds_w1', SAGE_russia_ds_w1)
#opal.symbol_import(o,'SAGE_russia_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_russia_ds_w1$ah) <- ds_label_all
ah <- SAGE_russia_ds_w1
save(ah, file = "../RData/russia_w1/ah.RData")
rm(ah)
```






## Mexico

### Wave 1 

#### Study-specific variable description

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | `991 = CAPI/interviewer error` |
| | `992 = Impute` |`992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable` | `996 = Not applicable` |
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` |`998 = Do not know` |
| | `999 = Missing` | `999 = Missing` |
| **Description** |  |  |


```{r assign_Mexico, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/mexico_w1/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SAGE/RData/mexico_w1/dbp.RData")
SAGE_mex_ah <- merge(sbp, dbp, by = "id")
```

```{r local_Mexico, echo=FALSE}
vari <- SAGE_mex_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_mex_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- SAGE_mex_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(SAGE_mex_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp>= 140 into 1`
  + `sbp< 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_Mexico}
SAGE_mex_ah$dh <- car::recode(SAGE_mex_ah$dbp, "0:89.99999999=0; 90:900=1")
SAGE_mex_ah$sh <- car::recode(SAGE_mex_ah$sbp, "0:139.9999999=0; 140:900=1")

SAGE_mexico_ds_w1 <- tibble(id=SAGE_mex_ah$id)
SAGE_mexico_ds_w1$ah <- NA
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==0 & SAGE_mex_ah$sh==0)] <- 0
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==1 & SAGE_mex_ah$sh==0)] <- 1
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==0 & SAGE_mex_ah$sh==1)] <- 1
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==1 & SAGE_mex_ah$sh==1)] <- 1
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==999 & SAGE_mex_ah$sh==999)] <- 999
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==998 & SAGE_mex_ah$sh==998)] <- 998
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==997 & SAGE_mex_ah$sh==997)] <- 997

SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==1 & SAGE_mex_ah$sh==998)] <- 1
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==998 & SAGE_mex_ah$sh==1)] <- 1
SAGE_mexico_ds_w1$ah[which(SAGE_mex_ah$dh==998 & SAGE_mex_ah$sh==0)] <- 998

SAGE_mexico_ds_w1$ah <- labelled(SAGE_mexico_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_Mexico, echo=FALSE}
vari <- SAGE_mexico_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SAGE_mexico_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_Mexico, echo=FALSE}
```

```{r importOPAL_Mexico, echo=FALSE}
#val_labels(SAGE_mexico_ds_w1) <- NULL
#opal.assign.data(o, 'SAGE_mexico_ds_w1', SAGE_mexico_ds_w1)
#opal.symbol_import(o,'SAGE_mexico_ds_w1', project='_Athlos_Harm_Dataset')
label(SAGE_mexico_ds_w1$ah) <- ds_label_all
ah <- SAGE_mexico_ds_w1
save(ah, file = "../RData/mexico_w1/ah.RData")
rm(ah)
```



# Quality estimation

No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
