==> SAGE/6_Physical_measures/13_heart_rate.Rmd <==
---
title: "Heart Rate"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/heart_rateDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/heart_rateDS.R')
```

# Data process





## South Africa

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_p`|`q2502a_p`|`q2503a_p`|
|-|-|-|-|
| **Label** |`q2501a: pulse (1)`|`q2502a: pulse (2)`|`q2503a: pulse (3)`|
| **Table name** |`southAfricaINDdata`|`southAfricaINDdata`|`southAfricaINDdata`|
| **Units** |`bpm`|`bpm`|`bpm`|
| **Missings** |`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|
| **Description** |  |  |  |




```{r assign SA_w1, echo=F}

opal.assign.table.tibble(o, 'SAGE_SA_w1','SAGE.southAfricaINDdata', variables=list('q2501a_p','q2502a_p','q2503a_p'), missings = TRUE)
SAGE_SA_w1 <- opal.execute(o,'SAGE_SA_w1')

```

```{r local SA_w1, echo=F}

vbl <- SAGE_SA_w1$q2501a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2501a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_SA_w1, aes(q2501a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_SA_w1$q2502a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2502a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_SA_w1, aes(q2502a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_SA_w1$q2503a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2503a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_SA_w1, aes(q2503a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables, they have to be recoded as follows:

* `average of the scores of q2501a_p,q2502a_p,q2503a_p into heart_rate`
* `-9 into 997`
* `-8 into 998`
* `NA into 999`

```{r harmo SA_w1}

SAGE_ds_SA_w1 <- tibble(id=SAGE_SA_w1$id)

SAGE_SA_w1$q2501a <- car::recode(SAGE_SA_w1$q2501a_p, "c(-9,-8) = NA")
SAGE_SA_w1$q2502a <- car::recode(SAGE_SA_w1$q2502a_p, "c(-9,-8) = NA")
SAGE_SA_w1$q2503a <- car::recode(SAGE_SA_w1$q2503a_p, "c(-9,-8) = NA")

SAGE_ds_SA_w1$heart_rate <- rowMeans(cbind(SAGE_SA_w1$q2501a,SAGE_SA_w1$q2502a,SAGE_SA_w1$q2503a),na.rm = T)

SAGE_ds_SA_w1$heart_rate[which(is.na(SAGE_ds_SA_w1$heart_rate) & (SAGE_SA_w1$q2501a_p == -9 | SAGE_SA_w1$q2502a_p == -9 | SAGE_SA_w1$q2503a_p == -9))] <- 997
SAGE_ds_SA_w1$heart_rate[which(is.na(SAGE_ds_SA_w1$heart_rate) & (SAGE_SA_w1$q2501a_p == -8 | SAGE_SA_w1$q2502a_p == -8 | SAGE_SA_w1$q2503a_p == -8))] <- 998

SAGE_ds_SA_w1$heart_rate <- car::recode(SAGE_ds_SA_w1$heart_rate, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript SA_w1, echo=F}

vbl <- SAGE_ds_SA_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_SA_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation SA_w1, echo=F}
```






## China

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_p`|`q2502a_p`|`q2503a_p`|
|-|-|-|-|
| **Label** |`q2501a: pulse (1)`|`q2502a: pulse (2)`|`q2503a: pulse (3)`|
| **Table name** |`chinaINDData`|`chinaINDData`|`chinaINDData`|
| **Units** |`bpm`|`bpm`|`bpm`|
| **Missings** |`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|
| **Description** |  |  |  |




```{r assign china_w1, echo=F}

opal.assign.table.tibble(o, 'SAGE_china_w1','SAGE.chinaINDData', variables=list('q2501a_p','q2502a_p','q2503a_p'), missings = TRUE)
SAGE_china_w1 <- opal.execute(o,'SAGE_china_w1')

```

```{r local china_w1, echo=F}

vbl <- SAGE_china_w1$q2501a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2501a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_china_w1, aes(q2501a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_china_w1$q2502a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2502a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_china_w1, aes(q2502a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_china_w1$q2503a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2503a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_china_w1, aes(q2503a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables, they have to be recoded as follows:

* `average of the scores of q2501a_p,q2502a_p,q2503a_p into heart_rate`
* `-9 into 997`
* `-8 into 998`
* `NA into 999`

```{r harmo china_w1}

SAGE_ds_china_w1 <- tibble(id=SAGE_china_w1$id)

SAGE_china_w1$q2501a <- car::recode(SAGE_china_w1$q2501a_p, "c(-9,-8) = NA")
SAGE_china_w1$q2502a <- car::recode(SAGE_china_w1$q2502a_p, "c(-9,-8) = NA")
SAGE_china_w1$q2503a <- car::recode(SAGE_china_w1$q2503a_p, "c(-9,-8) = NA")

SAGE_ds_china_w1$heart_rate <- rowMeans(cbind(SAGE_china_w1$q2501a,SAGE_china_w1$q2502a,SAGE_china_w1$q2503a),na.rm = T)

SAGE_ds_china_w1$heart_rate[which(is.na(SAGE_ds_china_w1$heart_rate) & (SAGE_china_w1$q2501a_p == -9 | SAGE_china_w1$q2502a_p == -9 | SAGE_china_w1$q2503a_p == -9))] <- 997
SAGE_ds_china_w1$heart_rate[which(is.na(SAGE_ds_china_w1$heart_rate) & (SAGE_china_w1$q2501a_p == -8 | SAGE_china_w1$q2502a_p == -8 | SAGE_china_w1$q2503a_p == -8))] <- 998

SAGE_ds_china_w1$heart_rate <- car::recode(SAGE_ds_china_w1$heart_rate, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript china_w1, echo=F}

vbl <- SAGE_ds_china_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_china_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation china_w1, echo=F}
```








## Ghana

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_p`|`q2502a_p`|`q2503a_p`|
|-|-|-|-|
| **Label** |`q2501a: pulse (1)`|`q2502a: pulse (2)`|`q2503a: pulse (3)`|
| **Table name** |`ghanaINDData`|`ghanaINDData`|`ghanaINDData`|
| **Units** |`bpm`|`bpm`|`bpm`|
| **Missings** |`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|
| **Description** |  |  |  |




```{r assign ghana_w1, echo=F}

opal.assign.table.tibble(o, 'SAGE_ghana_w1','SAGE.ghanaINDData', variables=list('q2501a_p','q2502a_p','q2503a_p'), missings = TRUE)
SAGE_ghana_w1 <- opal.execute(o,'SAGE_ghana_w1')

```

```{r local ghana_w1, echo=F}

vbl <- SAGE_ghana_w1$q2501a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2501a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_ghana_w1, aes(q2501a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_ghana_w1$q2502a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2502a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_ghana_w1, aes(q2502a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_ghana_w1$q2503a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2503a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_ghana_w1, aes(q2503a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables, they have to be recoded as follows:

* `average of the scores of q2501a_p,q2502a_p,q2503a_p into heart_rate`
* `-9 into 997`
* `-8 into 998`
* `NA into 999`

```{r harmo ghana_w1}

SAGE_ds_ghana_w1 <- tibble(id=SAGE_ghana_w1$id)

SAGE_ghana_w1$q2501a <- car::recode(SAGE_ghana_w1$q2501a_p, "c(-9,-8) = NA")
SAGE_ghana_w1$q2502a <- car::recode(SAGE_ghana_w1$q2502a_p, "c(-9,-8) = NA")
SAGE_ghana_w1$q2503a <- car::recode(SAGE_ghana_w1$q2503a_p, "c(-9,-8) = NA")

SAGE_ds_ghana_w1$heart_rate <- rowMeans(cbind(SAGE_ghana_w1$q2501a,SAGE_ghana_w1$q2502a,SAGE_ghana_w1$q2503a),na.rm = T)

SAGE_ds_ghana_w1$heart_rate[which(is.na(SAGE_ds_ghana_w1$heart_rate) & (SAGE_ghana_w1$q2501a_p == -9 | SAGE_ghana_w1$q2502a_p == -9 | SAGE_ghana_w1$q2503a_p == -9))] <- 997
SAGE_ds_ghana_w1$heart_rate[which(is.na(SAGE_ds_ghana_w1$heart_rate) & (SAGE_ghana_w1$q2501a_p == -8 | SAGE_ghana_w1$q2502a_p == -8 | SAGE_ghana_w1$q2503a_p == -8))] <- 998

SAGE_ds_ghana_w1$heart_rate <- car::recode(SAGE_ds_ghana_w1$heart_rate, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript ghana_w1, echo=F}

vbl <- SAGE_ds_ghana_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_ghana_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation ghana_w1, echo=F}
```






## India

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_p`|`q2502a_p`|`q2503a_p`|
|-|-|-|-|
| **Label** |`q2501a: pulse (1)`|`q2502a: pulse (2)`|`q2503a: pulse (3)`|
| **Table name** |`indiaINDData`|`indiaINDData`|`indiaINDData`|
| **Units** |`bpm`|`bpm`|`bpm`|
| **Missings** |`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|
| **Description** |  |  |  |




```{r assign india_w1, echo=F}

opal.assign.table.tibble(o, 'SAGE_india_w1','SAGE.indiaINDData', variables=list('q2501a_p','q2502a_p','q2503a_p'), missings = TRUE)
SAGE_india_w1 <- opal.execute(o,'SAGE_india_w1')

```

```{r local india_w1, echo=F}

vbl <- SAGE_india_w1$q2501a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2501a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_india_w1, aes(q2501a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_india_w1$q2502a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2502a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_india_w1, aes(q2502a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_india_w1$q2503a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2503a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_india_w1, aes(q2503a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables, they have to be recoded as follows:

* `average of the scores of q2501a_p,q2502a_p,q2503a_p into heart_rate`
* `-9 into 997`
* `-8 into 998`
* `NA into 999`

```{r harmo india_w1}

SAGE_ds_india_w1 <- tibble(id=SAGE_india_w1$id)

SAGE_india_w1$q2501a <- car::recode(SAGE_india_w1$q2501a_p, "c(-9,-8) = NA")
SAGE_india_w1$q2502a <- car::recode(SAGE_india_w1$q2502a_p, "c(-9,-8) = NA")
SAGE_india_w1$q2503a <- car::recode(SAGE_india_w1$q2503a_p, "c(-9,-8) = NA")

SAGE_ds_india_w1$heart_rate <- rowMeans(cbind(SAGE_india_w1$q2501a,SAGE_india_w1$q2502a,SAGE_india_w1$q2503a),na.rm = T)

SAGE_ds_india_w1$heart_rate[which(is.na(SAGE_ds_india_w1$heart_rate) & (SAGE_india_w1$q2501a_p == -9 | SAGE_india_w1$q2502a_p == -9 | SAGE_india_w1$q2503a_p == -9))] <- 997
SAGE_ds_india_w1$heart_rate[which(is.na(SAGE_ds_india_w1$heart_rate) & (SAGE_india_w1$q2501a_p == -8 | SAGE_india_w1$q2502a_p == -8 | SAGE_india_w1$q2503a_p == -8))] <- 998

SAGE_ds_india_w1$heart_rate <- car::recode(SAGE_ds_india_w1$heart_rate, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript india_w1, echo=F}

vbl <- SAGE_ds_india_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_india_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation india_w1, echo=F}
```






## Mexico

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_p`|`q2502a_p`|`q2503a_p`|
|-|-|-|-|
| **Label** |`q2501a: pulse (1)`|`q2502a: pulse (2)`|`q2503a: pulse (3)`|
| **Table name** |`mexicoINDdata`|`mexicoINDdata`|`mexicoINDdata`|
| **Units** |`bpm`|`bpm`|`bpm`|
| **Missings** |`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|`NA`|
| **Description** |  |  |  |




```{r assign mexico_w1, echo=F}

opal.assign.table.tibble(o, 'SAGE_mexico_w1','SAGE.mexicoINDdata', variables=list('q2501a_p','q2502a_p','q2503a_p'), missings = TRUE)
SAGE_mexico_w1 <- opal.execute(o,'SAGE_mexico_w1')

```

```{r local mexico_w1, echo=F}

vbl <- SAGE_mexico_w1$q2501a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2501a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_mexico_w1, aes(q2501a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_mexico_w1$q2502a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2502a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_mexico_w1, aes(q2502a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_mexico_w1$q2503a_p
vbl_miss <- NA
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2503a_p") 

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables, they have to be recoded as follows:

* `average of the scores of q2501a_p,q2502a_p,q2503a_p into heart_rate`
* `-9 into 997`
* `-8 into 998`
* `NA into 999`

```{r harmo mexico_w1}

SAGE_ds_mexico_w1 <- tibble(id=SAGE_mexico_w1$id)

SAGE_mexico_w1$q2501a <- car::recode(SAGE_mexico_w1$q2501a_p, "c(-9,-8) = NA")
SAGE_mexico_w1$q2502a <- car::recode(SAGE_mexico_w1$q2502a_p, "c(-9,-8) = NA")
SAGE_mexico_w1$q2503a <- car::recode(SAGE_mexico_w1$q2503a_p, "c(-9,-8) = NA")

SAGE_ds_mexico_w1$heart_rate <- rowMeans(cbind(SAGE_mexico_w1$q2501a,SAGE_mexico_w1$q2502a,SAGE_mexico_w1$q2503a),na.rm = T)

SAGE_ds_mexico_w1$heart_rate[which(is.na(SAGE_ds_mexico_w1$heart_rate) & (SAGE_mexico_w1$q2501a_p == -9 | SAGE_mexico_w1$q2502a_p == -9 | SAGE_mexico_w1$q2503a_p == -9))] <- 997
SAGE_ds_mexico_w1$heart_rate[which(is.na(SAGE_ds_mexico_w1$heart_rate) & (SAGE_mexico_w1$q2501a_p == -8 | SAGE_mexico_w1$q2502a_p == -8 | SAGE_mexico_w1$q2503a_p == -8))] <- 998

SAGE_ds_mexico_w1$heart_rate <- car::recode(SAGE_ds_mexico_w1$heart_rate, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript mexico_w1, echo=F}

vbl <- SAGE_ds_mexico_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_mexico_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation mexico_w1, echo=F}
```







## Russian Federation

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_p`|`q2502a_p`|`q2503a_p`|
|-|-|-|-|
| **Label** |`q2501a: pulse (1)`|`q2502a: pulse (2)`|`q2503a: pulse (3)`|
| **Table name** |`russiaINDdata`|`russiaINDdata`|`russiaINDdata`|
| **Units** |`bpm`|`bpm`|`bpm`|
| **Missings** |`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|`-9 = refused`<br/>`-8 = don't know`|
| **Description** |  |  |  |




```{r assign russia_w1, echo=F}

opal.assign.table.tibble(o, 'SAGE_russia_w1','SAGE.russiaINDdata', variables=list('q2501a_p','q2502a_p','q2503a_p'), missings = TRUE)
SAGE_russia_w1 <- opal.execute(o,'SAGE_russia_w1')

```

```{r local russia_w1, echo=F}

vbl <- SAGE_russia_w1$q2501a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2501a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_russia_w1, aes(q2501a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_russia_w1$q2502a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2502a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_russia_w1, aes(q2502a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

vbl <- SAGE_russia_w1$q2503a_p
vbl_miss <- c(-9,-8)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2503a_p") 
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SAGE_russia_w1, aes(q2503a_p)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables, they have to be recoded as follows:

* `average of the scores of q2501a_p,q2502a_p,q2503a_p into heart_rate`
* `-9 into 997`
* `-8 into 998`
* `NA into 999`

```{r harmo russia_w1}

SAGE_ds_russia_w1 <- tibble(id=SAGE_russia_w1$id)

SAGE_russia_w1$q2501a <- car::recode(SAGE_russia_w1$q2501a_p, "c(-9,-8) = NA")
SAGE_russia_w1$q2502a <- car::recode(SAGE_russia_w1$q2502a_p, "c(-9,-8) = NA")
SAGE_russia_w1$q2503a <- car::recode(SAGE_russia_w1$q2503a_p, "c(-9,-8) = NA")

SAGE_ds_russia_w1$heart_rate <- rowMeans(cbind(SAGE_russia_w1$q2501a,SAGE_russia_w1$q2502a,SAGE_russia_w1$q2503a),na.rm = T)

SAGE_ds_russia_w1$heart_rate[which(is.na(SAGE_ds_russia_w1$heart_rate) & (SAGE_russia_w1$q2501a_p == -9 | SAGE_russia_w1$q2502a_p == -9 | SAGE_russia_w1$q2503a_p == -9))] <- 997
SAGE_ds_russia_w1$heart_rate[which(is.na(SAGE_ds_russia_w1$heart_rate) & (SAGE_russia_w1$q2501a_p == -8 | SAGE_russia_w1$q2502a_p == -8 | SAGE_russia_w1$q2503a_p == -8))] <- 998

SAGE_ds_russia_w1$heart_rate <- car::recode(SAGE_ds_russia_w1$heart_rate, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript russia_w1, echo=F}

vbl <- SAGE_ds_russia_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SAGE_ds_russia_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation russia_w1, echo=F}
```








## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with continuous data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for continuous data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of column with summary table categories
  t.summ <- summary(m.hds[[1]][2])[1:6]
  # Adding of missing/no-missing values categories
  t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    # First, summary values
    t.summ <- summary(m.hds[[i]][2])[1:6]
    # Next, missing/no-missing values
    t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
  }
  # Add sample size for each wave/population
  t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))
  # Add wave/population names
  dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```





```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


sage.cw <- list(SA_w1 = c("SA_w1","South Africa w1"), ghana_w1 = c("ghana_w1","Ghana w1"), china_w1 = c("china_w1","China w1"), india_w1 = c("india_w1","India w1"), russia_w1 = c("russia_w1","Russian Federation w1"), mexico_w1 = c("mexico_w1","Mexico w1"))

# Consider only harmonised waves
l.hds <- lapply(sage.cw, function(wname) if(exists(paste0("SAGE_ds_",wname[1]))){wname = list(get(paste0("SAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```







# Quality estimation




<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
